Move MPSettings interfaces to DeviceSettings
