+++ src/Mooltipass/MPSettingsBLE.h
@@ -1,15 +1,18 @@
 #ifndef MPSETTINGSBLE_H
 #define MPSETTINGSBLE_H
 
-#include "MPSettings.h"
 #include "DeviceSettingsBLE.h"
 
-class MPSettingsBLE : public DeviceSettingsBLE, public MPSettings
+class MPDevice;
+class IMessageProtocol;
+
+class MPSettingsBLE : public DeviceSettingsBLE
 {
     Q_OBJECT
 
 public:
     MPSettingsBLE(MPDevice *parent, IMessageProtocol *mesProt);
+    virtual ~MPSettingsBLE() override = default;
 
     void loadParameters() override;
     void updateParam(MPParams::Param param, int val) override;
@@ -18,7 +21,10 @@ private slots:
     void setSettings();
 
 private:
-    void connectSendParams(WSServerCon* wsServerCon);
+    void connectSendParams(QObject* slotObject) override;
+
+    MPDevice* mpDevice = nullptr;
+    IMessageProtocol* pMesProt = nullptr;
 
     QByteArray m_lastDeviceSettings;
 };
