+++ src/Mooltipass/MPSettingsMini.h
@@ -1,19 +1,25 @@
 #ifndef MPSETTINGSMINI_H
 #define MPSETTINGSMINI_H
 
-#include "MPSettings.h"
 #include "DeviceSettingsMini.h"
 
-class MPSettingsMini : public DeviceSettingsMini, public MPSettings
+class MPDevice;
+class IMessageProtocol;
+
+class MPSettingsMini : public DeviceSettingsMini
 {
     Q_OBJECT
 
 public:
     MPSettingsMini(MPDevice *parent, IMessageProtocol *mesProt);
-    virtual ~MPSettingsMini() {}
+    virtual ~MPSettingsMini() override = default;
 
     void loadParameters() override;
     void updateParam(MPParams::Param param, int val) override;
+
+private:
+    MPDevice* mpDevice = nullptr;
+    IMessageProtocol* pMesProt = nullptr;
 };
 
 #endif // MPSETTINGSMINI_H
